dental-care-be/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.0.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==45.0.3
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
nodeenv==1.9.1
passlib==1.7.4
prisma==0.15.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.9.1
pydantic_core==2.27.2
python-dotenv==1.1.0
python-jose==3.5.0
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
tomlkit==0.13.3
typing-inspection==0.4.1
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
